fix(ports): keep statement evidence contracts out of adapters
Why:
- the shared statement extraction rollout left source adapters importing an application-layer batch type
- that broke the adapter boundary guard and failed the full quality gate pass

What:
- move the shared statement evidence batch contract to `tallylot.ports.evidence`
- keep the application statement extraction package re-exporting the batch type for callers that use the application seam
- update Binance and Shakepay adapters to depend on the port-level contract instead of importing from `application/`

Checks:
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run ruff check src/tallylot/ports/evidence.py src/tallylot/application/evidence/statement_extraction src/tallylot/adapters/sources/platforms/binance/adapter.py src/tallylot/adapters/sources/platforms/shakepay/adapter.py
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run mypy src/tallylot/ports/evidence.py src/tallylot/application/evidence/statement_extraction src/tallylot/adapters/sources/platforms/binance/adapter.py src/tallylot/adapters/sources/platforms/shakepay/adapter.py
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run pyright src/tallylot/ports/evidence.py src/tallylot/application/evidence/statement_extraction src/tallylot/adapters/sources/platforms/binance/adapter.py src/tallylot/adapters/sources/platforms/shakepay/adapter.py
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run pytest -q --no-cov tests/contract/test_standards_guards.py::test_adapter_production_modules_do_not_import_application_or_infrastructure src/tallylot/adapters/sources/platforms/binance/tests/test_normalize.py src/tallylot/adapters/sources/platforms/shakepay/tests/test_adapter.py tests/unit/application/normalization/test_normalization_service_hardening.py
- UV_PROJECT_ENVIRONMENT="$HOME/.venvs/tallylot-py312" uv run python -m tools.run_quality_gates --full-tests
