+++ src/tallylot/adapters/sources/platforms/shakepay/adapter.py
@@ -28,15 +28,15 @@ from tallylot.adapters.support import (
     skip_files_outside_profile_families,
 )
 from tallylot.adapters.support.drafts import symbol_claim, translation_batch_from_drafts
-from tallylot.application.evidence.statement_extraction import (
-    StatementBalanceEvidenceBatch,
-)
 from tallylot.domain.instruments import InstrumentIdentityClaim, InstrumentKind
 from tallylot.domain.issues import IssueRecord, NormalizationReviewRecord
 from tallylot.domain.reconciliation import BalanceEvidence
 from tallylot.domain.types import AdapterId, JsonValue
 from tallylot.ports.adapter_contracts import AdapterCapability, AdapterManifest
-from tallylot.ports.evidence import LocationInventoryRecord
+from tallylot.ports.evidence import (
+    LocationInventoryRecord,
+    StatementBalanceEvidenceBatch,
+)
 from tallylot.ports.intake_routing import (
     IntakeFileFacts,
     IntakeRoute,
